Implement in-memory encryption of "Protected" fields and attachments.
This reduces the risk of sensitive information being swapped to disk.
It is NOT a guarantee that no sensitive information can be extracted
from the process memory (we have other permission-based countermeasures
for that in place). Passwords and other data still need to be accessed
and shown in plaintext at some point and there is no way to do that
without having a decrypted copy in memory. However, this patch reduces
the risk of long-lived objects being swapped to disk in plain text from
where they may be recoverable by an attacker.

Encryption only applies to fields marked as "Protected" (which is the
default for the password field) and attachments. Attachments are
always encrypted, independently of whether they are marked as
"Protected", since there is no GUI option to set this flag and we
can safely assume that attachments very often contain private keys
and other highly critical information.